Move FuzzyDateField and SubmitButtonField to common.fields
