+++ common/fields/__init__.py
@@ -1,0 +1,2 @@
+from .fuzzy_date_field import FuzzyDateField  # noqa
+from .submit_button_field import SubmitButtonField, SubmitButtonInput  # noqa
