+++ productions/forms.py
@@ -1,11 +1,11 @@
 from django import forms
 from django.forms.formsets import formset_factory
 from django.forms.models import BaseModelFormSet, ModelFormOptions, inlineformset_factory
-from fuzzy_date_field import FuzzyDateField
 from nick_field import NickField
 
 from demoscene.forms.common import BaseExternalLinkFormSet, BaseTagsForm, ExternalLinkForm
 from demoscene.models import Edit
+from common.fields import FuzzyDateField
 from common.utils import groklinks
 from demoscene.utils.party_field import PartyField
 from platforms.models import Platform
